add project open API

+++ src/main/java/KU/GraduationProject/BasicServer/dto/projectDto/furnitureDto.java
@@ -11,8 +11,8 @@ public class furnitureDto {
 
     private String name;
 
-    private Long x;
+    private double x;
 
-    private Long y;
+    private double y;
 
 }
